mount.crypt: look for correct dm name
When pam_mount checks whether the encrypted filesystem is already
mounted, it looks for the device mapper name based on the device
pointed to (with '/' being replaced with '_'). But, if the encrypted
filesystem is a file, pointed to by a loop device, then pam_mount does
not find the previous mount and mounts a second time (and again for
each subsequent login). I have modified the device mapper name we are
creating to match the pattern of what the library is looking for. e.g.
encrypted filesystem in a file (/home/guest.fs); library looks for
"/dev/mapper/_home_guest.fs"; with old naming, the device mapper name
will be something like "/dev/mapper/_dev_loop0"