src: avoid garbage in readconfig.c's output
The issue is that readconfig.c's read_command callback does not put a
NULL into the next array entry after the last segment read, which
means, if the memory is not NULL (e.g. being recycled from the malloc
heap) it will end up with unexpected rubbish in the parameter list.