src: free pam structure only once
When pam_close_session is called twice, PMConfig will be cleaned twice.
By declaring it as a PAM item, it will only be freed once, on the
pam_end() call (which calls clean_config automatically). Additionally,
make the freeconfig() function reset the config so it can be called more
than once without crashing, just in case.