fixed: mount flag and value were one argument
SF bug id #1918174. The argument generation in pam_mount was done in
a suboptimal way. It used the %(before="-o " OPTIONS) placeholder,
which makes sure to only output -o when there are actually options to
print. The problem incurred is that the expanded result was of course
still one string, i.e. "-o bind", not "-o", "bind". Most mount
programs coped with it (seems to have required a getopt
implementation that supports squashing _and_ skips whitespace), but
the FUSE helper and sshfs apparently do not.

To fix this, a new placeholder %(ifnempty="-o" OPTIONS) was added in
libHX 1.15, and the pam_mount config file now reads

	%(ifnempty="-o" OPTIONS) %(OPTIONS)

so as to produce two distinct strings/arguments. (Note: empty
strings are not added to the argv array for the mount helper.)