src: destroy volumes on error
When there is a problem with a volume, the function rc_volume() just
returns, leaving the unfinished volume in the config->volume list.
Though it is cleaned up in freeconfig(), it probably is not nice for
do_mount() to attempt to mount it. Free it right when the problem
occurs.