src: traverse non-whitespace properly, check for '\0'
Checking for !isspace is not enough, because we would overrun
the '\0'. So check for its presence.

While this fixes something essential, it is not considered
security critical, as the analyzed textstrings always (unless
the kernel or cryptsetup has been compromised in which case you
alread lost) have the required amount of fields.