luserconf: re-enforce three-wall option checks
Denied and Required mount options were checked, but the result was
ignored (merge resolution error at commit
feddd7c0a49205102c48442da247a9ee2cf0edc6). However, because the list
of Allowed options was enforced, and all non-denied and Required
options had to be listed in the Allowed list. Therefore, only in the
presence of ambiguous configuration settings would it become a
problem, for example:

	<mntoptions allow="suid">
	<mntoptions deny="suid">