nucrypt: add missing return statements during loop+crypto setup
When no free loop devices were found, the missing return statements
cause control flow to go further than it should. While NULL is
passed to functions and whatnot, a NULL-dereference does not seem
to happen from reading the code - it will just fail because too few
arguments are passed to cryptsetup, and calling open(2) with NULL
produces -EFAULT on Linux.
Affected versions are 1.6 to 1.9.