config: broaden variable expansion to cover other option separators
Using an option string like "br:/tmp/%(USER)=rw" with AUFS resulted
in no expansion, because the key, assumed to be left of the equals
sign, was not expanded. However, AUFS uses a different option
notation, and so our splitting on ',' does not yield what would have
been expected. While that is a larger topic to discover, for now, let
pam_mount do expansion on the keys to make the above option string
expand to what users expected.