mtab: avoid NULL dereference when there is an empty line
In the occasion that a non-empty line is present in /etc/mtab
(somehow my system ends up having a line with a single space in it),
pam_mount segfaults while parsing /etc/mtab. This is caused by the
parser returning NULL-pointers, which are passed on to srcmp(). The
attached patch checks for non-NULL before passing.

References: http://sf.net/support/tracker.php?aid=2897206