log: use single quotes for arglist logging
Currently pam_mount prints, for example:

 [mount.crypt] [-ofsk_cipher=...] [/dev/mapper/abc] [/home/abc]

This is a little awkward to copy-n-paste to the terminal, as the
shell does not see [/] as delimiters. Instead, let's have pam_mount
output it in single quotes

 'mount.crypt' '-ofsk_cipher=...' '/dev/mapper/abc' '/home/abc'

while that does not magically escape quotes inside the argument, it
should be sufficient for the developer, and definitely is
copypastable.