pmt-ehd: avoid needless overtruncation/sparsifying
The O_TRUNC flag would cause the file to be truncated to zero; the
subsequent truncate(2) call would then enlarge it again - with sparse
blocks this time. If the file already existed, that is a dull thing
to do, as the creator might have had in mind to use a preallocated
file. Just directly truncate to the right size, avoiding any
sparseing.