pam_mount: up the refcount once freeconfig is live
It seems that on Fedora, cleanconfig is called at the end of
session_open already, rather than at the end of session_close. Move the
refcount increment directly after telling PAM to use the "cleanconfig"
function, so that we do not run into a case where the refcount goes
below zero.

References: https://bugzilla.redhat.com/622930