pam_mount: restore group initialization
With the master Git pam-mount the user groups are not correctly
initialized, which leads to a login failure in my setup using sshfs
and fuse with a limitation on the allowed groups.

The problem is in initgroups2() where the malloc() return value is not
correctly checked, and this patchAA does the trick for me.