pam_mount: use getgrouplist to implement sgrp match
When using <volume sgrp="..."> (or <sgrp>), pam_mount would attempt to
retrieve the group's member list and check for the user's presence.
Some LDAP setups however can have a limit as to how many members are
returned.

Instead, ask for the user's group memberships, and check for the
group's presence. LDAP setups won't be having a limit on this (at
least not a larger one than the basic system), as it would impact all
kinds of permission checks.